Extract map overlay resolution out of router into map_overlay module
