+++ app/src-tauri/src/lib.rs
@@ -13,6 +13,7 @@ mod audio;
 mod commands;
 mod hotkeys;
 mod logging;
+mod map_overlay;
 pub mod overlay;
 mod router;
 pub mod service;
@@ -181,7 +182,7 @@ pub fn run() {
                     )
                     .ok()
                     .filter(|p| p.exists());
-                router::init_bundled_map_dir(bundled_map_dir);
+                map_overlay::init_bundled_map_dir(bundled_map_dir);
 
                 // Spawn the overlay update router (needs service handle for registry updates)
                 spawn_overlay_router(
